requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-colorfield==0.11.0
django-unfold==0.37.0
idna==3.10
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
stripe==10.12.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
